[+] sync maintaining angular 2 across client, desktop, and mobile [+] better logging for auth token failures [+] lodash and rsvp added for async series [+] `flag` moved to `designer` [+] socket service updated to notify only on server signals

+++ server_restapi/source/database/function/sp_accountToken_fetch_account.sql
@@ -9,7 +9,7 @@ DROP PROCEDURE IF EXISTS sp_accountToken_fetch_account;
 
 CREATE PROCEDURE sp_accountToken_fetch_account
 (
-    p_accountId INT(10)
+    p_accountId         BIGINT
 )
 this_proc:BEGIN
